more Room adoption - accounts@100%, some profiles/transactions

+++ app/src/main/java/net/ktnx/mobileledger/db/AccountValue.java
@@ -44,7 +44,7 @@ public class AccountValue {
     @ColumnInfo
     private float value;
     @ColumnInfo(defaultValue = "0")
-    private int generation = 0;
+    private long generation = 0;
     public long getId() {
         return id;
     }
@@ -70,10 +70,10 @@ public class AccountValue {
     public void setValue(float value) {
         this.value = value;
     }
-    public int getGeneration() {
+    public long getGeneration() {
         return generation;
     }
-    public void setGeneration(int generation) {
+    public void setGeneration(long generation) {
         this.generation = generation;
     }
 }
